stack based function implemented, functions transferred to the astar files (algorithms and tests), doc_strings added
